requirements.txt
arrow>=1.2.3
dj-database-url>=0.5.0
python-dotenv>=1.0.0
django-cors-headers>=4.4.0
Django>=4.2.1
djangorestframework>=3.15.1
django-allauth>=0.64.0
dj-rest-auth>=7.0.1
gunicorn>=21.2.0
pylibmc>=1.6.3
psycopg2-binary>=2.9.9
whitenoise>=6.5.0
django-cache-url>=3.4.4
simyan>=1.5.1
requests>=2.32.3
python-dateutil>=2.9.0.post0
safety
black>=24.8.0